instance functions should be in a file of the same name
